+++ temba/orgs/tests/test_org.py
@@ -409,13 +409,17 @@ class OrgTest(TembaTest):
         self.assertEqual(self.org.get_limit(Org.LIMIT_FIELDS), 250)
         self.assertEqual(self.org.get_limit(Org.LIMIT_GROUPS), 250)
         self.assertEqual(self.org.get_limit(Org.LIMIT_GLOBALS), 250)
+        self.assertEqual(self.org.get_limit(Org.LIMIT_CONTACTS), 50_000_000)
+        self.assertEqual(self.org.get_limit(Org.LIMIT_FLOWS), 10_000)
 
-        self.org.limits = dict(fields=500, groups=500)
+        self.org.limits = dict(fields=500, groups=500, contacts=100_000)
         self.org.save()
 
         self.assertEqual(self.org.get_limit(Org.LIMIT_FIELDS), 500)
         self.assertEqual(self.org.get_limit(Org.LIMIT_GROUPS), 500)
         self.assertEqual(self.org.get_limit(Org.LIMIT_GLOBALS), 250)
+        self.assertEqual(self.org.get_limit(Org.LIMIT_CONTACTS), 100_000)
+        self.assertEqual(self.org.get_limit(Org.LIMIT_FLOWS), 10_000)
 
     def test_org_api_rates(self):
         self.assertEqual(self.org.api_rates, {})